Initialized Estimate workflow. WIP: add student assignment with asset assignment to ensembles domain. WIP: Migrate from log to ses mailing. WIP: Build link to process verification action when link is clicked. WIP: testing for links to subordinate pages. WIP: Policy for director viewing of student pages to only allow teacher to view student. WIP: Policy for director viewing of student pages to prohibit non-verified email of director to view student entered information.

+++ app/Services/ConvertToUsdService.php
@@ -4,7 +4,7 @@ namespace App\Services;
 
 class ConvertToUsdService
 {
-    static public function penniesToUsd($pennies): int
+    static public function penniesToUsd($pennies): float
     {
         return number_format(($pennies / 100), 2);
     }
